+++ lib/onfido/resources/applicant.rb
@@ -4,6 +4,14 @@ module Onfido
       post(url: url_for('applicants'), payload: payload)
     end
 
+    def update(applicant_id, payload)
+      put(url: url_for("applicants/#{applicant_id}"), payload: payload)
+    end
+
+    def destroy(applicant_id)
+      delete(url: url_for("applicants/#{applicant_id}"))
+    end
+
     def find(applicant_id)
       get(url: url_for("applicants/#{applicant_id}"))
     end