+++ lib/onfido/resources/document.rb
@@ -9,5 +9,17 @@ module Onfido
         payload: payload
       )
     end
+
+    def find(applicant_id, document_id)
+      get(url: url_for("applicants/#{applicant_id}/documents/#{document_id}"))
+    end
+
+    def download(applicant_id, document_id)
+      get(url: url_for("applicants/#{applicant_id}/documents/#{document_id}/download"))
+    end
+
+    def all(applicant_id)
+      get(url: url_for("applicants/#{applicant_id}/documents"))
+    end
   end
 end