+++ spec/integrations/report_spec.rb
@@ -17,4 +17,24 @@ describe Onfido::Report do
       expect(response['reports'].count).to eq(2)
     end
   end
+
+  describe '#resume' do
+    let(:report_id) { '6951786-123123-422221' }
+    let(:check_id) { '1212121-123123-422221' }
+
+    it 'returns a success response' do
+      expect { report.resume(check_id, report_id) }.not_to raise_error
+    end
+  end
+
+  describe '#cancel' do
+    let(:report_id) { '6951786-123123-422221' }
+    let(:check_id) { '1212121-123123-422221' }
+
+    it 'returns a success response' do
+      expect { report.cancel(check_id, report_id) }.not_to raise_error
+    end
+  end
+
+
 end